requirements.txt
streamlit
PyPDF2
gTTS
SpeechRecognition
audio-recorder-streamlit
matplotlib
seaborn
plotly
wordcloud
networkx
reportlab
Pillow
pandas
google-generativeai
